+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 0.0.4 - in progress
-- 
+- Remove complement magic method, and add difference.
 
 0.0.3
 - Change to more fluent SDK: `select_TARGET(where='SOURCE', has='CRITERIA', ...)`